MMCA-4802 | Added CustomsDataStoreConnector for email retrieval
