+++ app/connectors/CustomsFinancialsApiConnector.scala
@@ -18,7 +18,7 @@ package connectors
 
 import config.AppConfig
 import models.request.{GuaranteeTransactionsRequest, IdentifierRequest}
-import models.{GuaranteeAccount, GuaranteeTransaction, RequestDates, EmailUnverifiedResponse}
+import models.{GuaranteeAccount, GuaranteeTransaction, RequestDates}
 import org.slf4j.LoggerFactory
 import play.api.mvc.AnyContent
 import repositories.CacheRepository
@@ -86,17 +86,12 @@ class CustomsFinancialsApiConnector @Inject()(httpClient: HttpClient,
     }
   }
 
-  def isEmailUnverified(implicit hc: HeaderCarrier): Future[Option[String]] = {
-    httpClient.GET[EmailUnverifiedResponse](
-      appConfig.customsFinancialsApi + "/subscriptions/unverified-email-display").map( res => res.unVerifiedEmail)
-  }
-
   def retrieveRequestedGuaranteeTransactionsDetail(
                                                     gan: String,
                                                     onlyOpenItems: Boolean,
                                                     from: LocalDate,
                                                     to: LocalDate)(
-    implicit hc: HeaderCarrier): Future[Either[GuaranteeResponses, Seq[GuaranteeTransaction]]] = {
+                                                    implicit hc: HeaderCarrier): Future[Either[GuaranteeResponses, Seq[GuaranteeTransaction]]] = {
 
     val openGuaranteeTransactionsRequest = GuaranteeTransactionsRequest(
       gan, onlyOpenItems, Some(RequestDates(from, to)))
