+++ test/connectors/CustomsDataStoreConnectorSpec.scala
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package connectors
+
+import models._
+import play.api.Application
+import play.api.inject.bind
+import play.api.test.Helpers._
+import uk.gov.hmrc.http.SessionId
+import uk.gov.hmrc.http.{HeaderCarrier, HttpClient}
+import utils.SpecBase
+
+import scala.concurrent.Future
+
+class CustomsDataStoreConnectorSpec extends SpecBase {
+
+  "isEmailUnverified" must {
+
+    "return unverified email" in new Setup {
+      when[Future[EmailUnverifiedResponse]](mockHttpClient.GET(any, any, any)(any, any, any))
+        .thenReturn(Future.successful(EmailUnverifiedResponse(Some("unverified@email.com"))))
+
+      val app: Application = application
+        .overrides(
+          bind[HttpClient].toInstance(mockHttpClient)
+        ).build()
+
+      val connector: CustomsDataStoreConnector = app.injector.instanceOf[CustomsDataStoreConnector]
+
+      running(app) {
+        val result = await(connector.isEmailUnverified(hc))
+        result mustBe Some("unverified@email.com")
+      }
+    }
+  }
+
+  trait Setup {
+    val sessionId: SessionId = SessionId("session_1234")
+    implicit val hc: HeaderCarrier = HeaderCarrier(sessionId = Some(sessionId))
+    val mockHttpClient: HttpClient = mock[HttpClient]
+  }
+}
